+++ Caper/Caper/SAMMappingUtilities.h
@@ -6,7 +6,7 @@ class SAMMappingUtilities :
 {
 public:
 	string GetSequence( string & aLine );
-  int GetIndex( string & aLine );    
+  long long GetIndex( string & aLine );    
   string GetName( string & aLine );
   string GetContigIdent( string & aLine );	
   string GetStrand( string & aLine );
